// features/SolicitudesActividades/models/Solicitud.ts
export type EstadoInvitacionDB = 'SENT' | 'SEEN' | 'MODIFIED' | 'MODIFIED_BY_HOST' | 'ACCEPTED_BY_HOST' | 'ACCEPTED' | 'REJECTED';
export type EstadoInvitacionUI = 'Pendiente' | 'Visto' | 'Modificado' | 'Modificado por creador' |'Aceptado por creador' | 'Aceptado' | 'Rechazado';

// Mapeo de estados de DB a UI
export const estadoInvitacionMapping: Record<EstadoInvitacionDB, EstadoInvitacionUI> = {
  'SENT': 'Pendiente',
  'SEEN': 'Visto',
  'MODIFIED': 'Modificado',
  'MODIFIED_BY_HOST': 'Modificado por creador',
  'ACCEPTED_BY_HOST': 'Aceptado por creador',
  'ACCEPTED': 'Aceptado',
  'REJECTED': 'Rechazado'
};
// ...
export interface Solicitud {
  solicitud_id: number;
  titulo: string;
  descripcion: string;
  created_by: number;
  fecha_inicio: string; // ISO 8601 UTC
  fecha_fin: string; // ISO 8601 UTC
  estado: EstadoInvitacionDB; // Estado de la invitación
  nombre: string; // Opcional: nombre del creador de la solicitud
  apellido: string; // Opcional: apellido del creador de la solicitud
}